+++ packages/quantic/cypress/e2e/facets-1/facet/facet.cypress.ts
@@ -2,7 +2,7 @@ import {performSearch} from '../../../page-objects/actions/action-perform-search
 import {configure} from '../../../page-objects/configurator';
 import {
   extractFacetValues,
-  getAlias,
+  getQueryAlias,
   InterceptAliases,
   interceptSearch,
   interceptSearchIndefinitely,
@@ -53,7 +53,7 @@ describe('Facet Test Suite', () => {
       performSearch();
     }
     if (waitForSearch) {
-      cy.wait(getAlias(options.useCase));
+      cy.wait(getQueryAlias(options.useCase));
     }
   }
 
@@ -68,7 +68,7 @@ describe('Facet Test Suite', () => {
   }
 
   function aliasFacetValues(useCase: string) {
-    cy.wait(getAlias(useCase)).then((interception) => {
+    cy.wait(getQueryAlias(useCase)).then((interception) => {
       const indexValues = extractFacetValues(interception.response);
       cy.wrap(indexValues).as(indexFacetValuesAlias.substring(1));
     });
@@ -191,7 +191,7 @@ describe('Facet Test Suite', () => {
 
             function collapseFacet() {
               Actions.clickCollapseButton();
-              cy.wait(getAlias(param.useCase));
+              cy.wait(getQueryAlias(param.useCase));
             }
 
             selectFirstFacetValue();
@@ -211,7 +211,7 @@ describe('Facet Test Suite', () => {
               function clearSelectedValues() {
                 Actions.clickExpandButton();
                 Actions.clickClearFilter();
-                cy.wait(getAlias(param.useCase));
+                cy.wait(getQueryAlias(param.useCase));
               }
               clearSelectedValues();
 
@@ -492,7 +492,7 @@ describe('Facet Test Suite', () => {
               function clearSelectedValues() {
                 Actions.clickExpandButton();
                 Actions.clickClearFilter();
-                cy.wait(getAlias(param.useCase));
+                cy.wait(getQueryAlias(param.useCase));
               }
 
               clearSelectedValues();
@@ -706,7 +706,7 @@ describe('Facet Test Suite', () => {
               scope('when clicking show less button', () => {
                 function showLessValues() {
                   showMoreValuesAgain();
-                  cy.wait(getAlias(param.useCase));
+                  cy.wait(getQueryAlias(param.useCase));
                   Actions.clickShowLessButton();
                 }
 
@@ -779,7 +779,7 @@ describe('Facet Test Suite', () => {
                 },
                 false
               );
-              cy.wait(getAlias(param.useCase)).then((interception) => {
+              cy.wait(getQueryAlias(param.useCase)).then((interception) => {
                 const facetRequest = interception.request.body.facets[0];
                 expect(facetRequest.sortCriteria).to.eq(sorting);
               });